+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Snippets: `devproxy-plugin-api-center-minimal-permissions` - ApiCenterMinimalPermissionsPlugin instance
 - Snippets: `devproxy-plugin-api-center-minimal-permissions-config` - ApiCenterMinimalPermissionsPlugin config section
+- Snippets: `devproxy-plugin-http-file-generator` - HttpFileGeneratorPlugin instance
+- Snippets: `devproxy-plugin-http-file-generator-config` - HttpFileGeneratorPlugin config section
 
 ### Changed: